Remove unnecessary casts to usize

+++ src/app_state.rs
@@ -338,11 +338,11 @@ impl TereAppState {
     /// Convert a cursor position (in the range 0..window_height) to an index
     /// into the currently visible items.
     pub fn cursor_pos_to_visible_item_index(&self, cursor_pos: usize) -> usize {
-        (cursor_pos + self.scroll_pos) as usize
+        cursor_pos + self.scroll_pos
     }
 
     pub fn get_item_at_cursor_pos(&self, cursor_pos: usize) -> Option<&CustomDirEntry> {
-        let idx = self.cursor_pos_to_visible_item_index(cursor_pos) as usize;
+        let idx = self.cursor_pos_to_visible_item_index(cursor_pos);
         self.visible_items().get(idx).copied()
     }
 
@@ -361,7 +361,7 @@ impl TereAppState {
     }
 
     pub fn get_match_locations_at_cursor_pos(&self, cursor_pos: usize) -> Option<&MatchesLocType> {
-        let idx = self.cursor_pos_to_visible_item_index(cursor_pos) as usize;
+        let idx = self.cursor_pos_to_visible_item_index(cursor_pos);
         if self.settings().filter_search {
             // NOTE: we assume that the matches is a sorted map
             self.ls_output_buf.matches.values().nth(idx)
